Create Data_NOAA_API.py
Here is how we will make the API call to pull in the data using NOAA. The class takes in a year, season, and location it will then build out the API call to pull the data it. Still need to make the named tuple section that will hold the data to be used.